\name{getQueryResults}
\alias{getQueryResults}
\alias{makeContext}
\alias{makeDynamicContext}
\alias{xmlTextNode}
\alias{zorbaPOSIXct}
\alias{StaticContext-class}
\alias{zorba::StaticContext-class}
\title{Functions for manipulating Zorba objects}
\description{}
\arguments{
}
\value{}
\examples{}
\keyword{programming}
