\name{XQuery-class}
\alias{zorba::XQuery-class}

\alias{getDocLoadingTime}
\alias{getDocLoadingTime,zorba::XQuery-method}
\alias{status}
\alias{status,zorba::XQuery-method}
\alias{setTimeout}
\alias{setTimeout,zorba::XQuery-method}

\alias{setDefaultCollection}


\title{Methods for compiled XQuery}
\description{
  These functions and the zorba::XQuery class provide access to the 
  methods in the corresponding C++ class.
  These allow us to query the state of the query along with
  functions for evaluating and getting the results from the query.
}
\usage{
}

\value{
}

\references{
}
\author{Duncan Temple Lang}
\seealso{
  \code{\link{xquery}}

  \code{\link{printPlan}}
}

\examples{

}
\keyword{IO}
