\name{zorbaVersion}
\alias{zorbaVersion}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Get Zorba version information}
\description{
  This function allows us to get the version information
  from the Zorba installation we are using.
}
\usage{
zorbaVersion()
}
\value{
 An named integer vector with 3 elements: major, minor and patch.
}
\references{
\url{http://www.zorba-xquery.com}
}
\author{
Duncan Temple Lang
}

\seealso{
 \code{\link{zorba}}
}
\examples{
}
\keyword{IO}

